refactor: moves mock enums
